Started lock-down of member controller. Hid some fields from non-admin members in the member::edit.

+++ app/Model/Member.php
@@ -1,4 +1,7 @@
 <?php
+
+	App::uses('AppModel', 'Model');
+
 	class Member extends AppModel {
 
 		public $primaryKey = 'member_id';
@@ -60,20 +63,24 @@
 
 
 		public function beforeSave($options = array()) {
-			# Have to do a few things before we save
-			$memberWillBeCurrentMember = $this->data['Member']['member_status'] == 2;
 
-			if( isset( $this->data['Member'] ) &&
-				isset( $this->data['Member']['member_number'] ) === false &&
-				$memberWillBeCurrentMember)
+			if( isset($this->data['Member']['member_status']) )
 			{
-				# We're setting this member to be a 'current member' for the first time, need to modify some things
+				# Have to do a few things before we save
+				$memberWillBeCurrentMember = $this->data['Member']['member_status'] == 2;
 
-				# Set the member number and join date
-				# Member number is totally fucked up with hard-coded entries and missing entries, so we need to find what the highest number is
-				$highestMemberNumber = $this->find( 'first', array( 'conditions' => array( 'Member.member_number !=' => null),  'order' => 'Member.member_number DESC', 'fields' => 'Member.member_number' ) );
-				$this->data['Member']['member_number'] = $highestMemberNumber['Member']['member_number'] + 1;
-				$this->data['Member']['join_date'] = date( 'Y-m-d' );
+				if( isset( $this->data['Member'] ) &&
+					isset( $this->data['Member']['member_number'] ) === false &&
+					$memberWillBeCurrentMember)
+				{
+					# We're setting this member to be a 'current member' for the first time, need to modify some things
+
+					# Set the member number and join date
+					# Member number is totally fucked up with hard-coded entries and missing entries, so we need to find what the highest number is
+					$highestMemberNumber = $this->find( 'first', array( 'conditions' => array( 'Member.member_number !=' => null),  'order' => 'Member.member_number DESC', 'fields' => 'Member.member_number' ) );
+					$this->data['Member']['member_number'] = $highestMemberNumber['Member']['member_number'] + 1;
+					$this->data['Member']['join_date'] = date( 'Y-m-d' );
+				}
 			}
 
 			# Must never ever ever alter the balance
@@ -83,30 +90,37 @@
 		}
 
 		public function clearGroupsIfMembershipRevoked($id, $newData) {
-			# If membership is being revoked, clear all groups
-			if($newData['Member']['member_status'] == 3)
+			if( isset($newData['Member']['member_status']) )
 			{
-				$this->MemberGroup->deleteAll(array( 'MemberGroup.member_id' => $id ));
+				# If membership is being revoked, clear all groups
+				if($newData['Member']['member_status'] == 3)
+				{
+					$this->MemberGroup->deleteAll(array( 'MemberGroup.member_id' => $id ));
+				}
 			}
 		}
 
 		public function addToCurrentMemberGroupIfStatusIsCurrentMember($id, $newData) {
-			# If membership is current_member, add to the current member group
-			if($newData['Member']['member_status'] == 2)
+
+			if( isset($newData['Member']['member_status']) )
 			{
-				$this->MemberGroup->deleteAll(array( 'MemberGroup.member_id' => $id, 'MemberGroup.grp_id' => 2 ));
-
-				# Group 2 is for current members
-				$currentGroups = Hash::extract($newData, 'Group.Group.{n}');
-				print_r($currentGroups);
-				if( in_array(2, $currentGroups) == false )
+				# If membership is current_member, add to the current member group
+				if($newData['Member']['member_status'] == 2)
 				{
-					array_push($currentGroups, array( 'grp_id' => 2, 'member_id' => $id ));
+					$this->MemberGroup->deleteAll(array( 'MemberGroup.member_id' => $id, 'MemberGroup.grp_id' => 2 ));
+
+					# Group 2 is for current members
+					$currentGroups = Hash::extract($newData, 'Group.Group.{n}');
+					print_r($currentGroups);
+					if( in_array(2, $currentGroups) == false )
+					{
+						array_push($currentGroups, array( 'grp_id' => 2, 'member_id' => $id ));
+					}
+
+					$newData['Group']['Group'] = $currentGroups;
+
+					$this->save($newData);
 				}
-
-				$newData['Group']['Group'] = $currentGroups;
-
-				$this->save($newData);
 			}
 		}
 
@@ -114,6 +128,27 @@
 		public function memberInGroup($memberId, $groupId)
 		{
 			return in_array($groupId, Hash::extract( $this->find('first', array( 'conditions' => array( 'Member.member_id' => $memberId ) ) ), 'Group.{n}.grp_id' ));
+		}
+
+		public static function isInGroup($user, $groupId)
+		{
+			if(	isset($user) &&
+				isset($user['Group']))
+			{
+				return in_array($groupId, Hash::extract($user['Group'], '{n}.grp_id'));
+			}
+
+			return false;
+		}
+
+		public static function isInGroupFullAccess($user)
+		{
+			return Member::isInGroup($user, 1);
+		}
+
+		public static function isInGroupMemberAdmin($user)
+		{
+			return Member::isInGroup($user, 5);
 		}
 	}
 ?>
